+++ lib/src/generators/json_serializable_generator.dart
@@ -20,14 +20,19 @@ class JsonSerializableGenerator extends ClassGeneratorStrategy {
     String className,
     Map<String, FieldInfo> fields,
     String Function(String jsonKey, Map<String, dynamic> schema) fromJsonExpression,
-    String Function(String fieldName, Map<String, dynamic> schema) toJsonExpression,
-  ) {
+    String Function(String fieldName, Map<String, dynamic> schema) toJsonExpression, {
+    bool useJsonKey = false,
+  }) {
     final buffer = StringBuffer()
       ..writeln('@JsonSerializable()')
       ..writeln('class $className {');
 
     // Поля
     fields.forEach((propName, field) {
+      // Генерируем @JsonKey, если нужно
+      if (useJsonKey && field.needsJsonKey(true)) {
+        buffer.writeln('  @JsonKey(name: \'${field.jsonKey}\')');
+      }
       buffer.writeln('  final ${field.dartType} ${field.camelCaseName};');
     });